refactor(grn): migrate components to feature directory and extract GoodsReceiptTable (Phase 6.1e)
- Move 'GoodsReceiptRow' from shared resources to 'features/grn/components'.
- Extract 'GoodsReceiptTable' component to encapsulate list rendering and filters.
- Update goods receipts list page to use new feature-specific components.
