+++ app/users/dashboard/resources/goods-receipts/page.tsx
@@ -2,27 +2,10 @@
 
 import { useState, useMemo } from 'react';
 import Link from 'next/link';
-import { useRouter } from 'next/navigation';
 import { routes } from '@/nav';
-import { Card, CardContent, CardHeader } from '@/components/shadcn/card';
+import { Card } from '@/components/shadcn/card';
 import { Button } from '@/components/shadcn/button';
-import { Input } from '@/components/shadcn/input';
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/shadcn/select';
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from '@/components/shadcn/table';
-import { Pagination, PageHeader } from '@/components/common';
+import { PageHeader } from '@/components/common';
 import {
   Plus,
   Loader2,
@@ -30,43 +13,17 @@ import {
   CalendarDays,
   IndianRupee,
   ShoppingCart,
-  Search,
 } from 'lucide-react';
-import {
-  Empty,
-  EmptyMedia,
-  EmptyHeader,
-  EmptyTitle,
-  EmptyDescription,
-} from '@/components/shadcn/empty';
-import { format } from 'date-fns';
 import { useGRNs } from '@/hooks/grn';
-import { GoodsReceiptRow } from '@/features/resources/components';
-
-const PAGE_SIZE_OPTIONS = [5, 10, 20, 50, 100];
+import { GoodsReceiptTable } from '@/features/grn/components';
 
 export default function GoodsReceiptsPage() {
-  const router = useRouter();
   const { data: grns = [], isLoading } = useGRNs();
 
   const [searchQuery, setSearchQuery] = useState('');
   const [projectFilter, setProjectFilter] = useState('all');
   const [currentPage, setCurrentPage] = useState(1);
   const [itemsPerPage, setItemsPerPage] = useState(10);
-  const [prevFilters, setPrevFilters] = useState({
-    searchQuery,
-    projectFilter,
-    itemsPerPage,
-  });
-
-  if (
-    prevFilters.searchQuery !== searchQuery ||
-    prevFilters.projectFilter !== projectFilter ||
-    prevFilters.itemsPerPage !== itemsPerPage
-  ) {
-    setPrevFilters({ searchQuery, projectFilter, itemsPerPage });
-    setCurrentPage(1);
-  }
 
   const filtered = useMemo(() => {
     const q = searchQuery.toLowerCase();
@@ -91,8 +48,7 @@ export default function GoodsReceiptsPage() {
 
   const totalPages = Math.ceil(filtered.length / itemsPerPage);
   const startIndex = (currentPage - 1) * itemsPerPage;
-  const endIndex = Math.min(startIndex + itemsPerPage, filtered.length);
-  const paginated = filtered.slice(startIndex, endIndex);
+  const paginated = filtered.slice(startIndex, startIndex + itemsPerPage);
 
   const now = new Date();
   const thisMonth = grns.filter((g) => {
@@ -129,7 +85,6 @@ export default function GoodsReceiptsPage() {
         }
       />
 
-      {/* Stats */}
       <Card className="gap-0 p-6">
         <div className="sm:divide-border grid grid-cols-2 gap-3 sm:grid-cols-4 sm:gap-0 sm:divide-x">
           <div className="flex flex-col gap-1 rounded-lg p-3 sm:rounded-none sm:pr-6">
@@ -199,132 +154,31 @@ export default function GoodsReceiptsPage() {
         </div>
       </Card>
 
-      {/* Table Card */}
-      <Card>
-        <CardHeader className="flex flex-row items-center gap-3 border-b px-4 py-1">
-          <div className="relative w-full max-w-xs">
-            <Search className="absolute top-1/2 left-3 size-3.5 -translate-y-1/2 text-zinc-400" />
-            <Input
-              value={searchQuery}
-              onChange={(e) => {
-                setSearchQuery(e.target.value);
-                setCurrentPage(1);
-              }}
-              placeholder="Search by GRN number, vendor, PO or project…"
-              className="h-8 pl-8 text-sm"
-            />
-          </div>
-          <Select
-            value={projectFilter}
-            onValueChange={(v) => {
-              setProjectFilter(v);
-              setCurrentPage(1);
-            }}
-          >
-            <SelectTrigger className="h-8 w-36 text-xs">
-              <SelectValue placeholder="Project" />
-            </SelectTrigger>
-            <SelectContent>
-              <SelectItem value="all">All Projects</SelectItem>
-              {projectOptions.map((p) => (
-                <SelectItem key={p} value={p}>
-                  {p}
-                </SelectItem>
-              ))}
-            </SelectContent>
-          </Select>
-          <div className="ml-auto flex items-center gap-2 border-l pl-3">
-            <span className="text-xs whitespace-nowrap text-zinc-500">
-              Rows per page
-            </span>
-            <Select
-              value={String(itemsPerPage)}
-              onValueChange={(v) => {
-                setItemsPerPage(Number(v));
-                setCurrentPage(1);
-              }}
-            >
-              <SelectTrigger className="h-8 w-16 text-xs">
-                <SelectValue />
-              </SelectTrigger>
-              <SelectContent>
-                {PAGE_SIZE_OPTIONS.map((n) => (
-                  <SelectItem key={n} value={String(n)}>
-                    {n}
-                  </SelectItem>
-                ))}
-              </SelectContent>
-            </Select>
-          </div>
-        </CardHeader>
-
-        {paginated.length > 0 ? (
-          <>
-            <CardContent className="p-0">
-              <Table>
-                <TableHeader>
-                  <TableRow>
-                    <TableHead className="pl-6">GRN Number</TableHead>
-                    <TableHead>Vendor</TableHead>
-                    <TableHead>Received On</TableHead>
-                    <TableHead>Purchase Order</TableHead>
-                    <TableHead>Project</TableHead>
-                    <TableHead>Items</TableHead>
-                    <TableHead className="pr-6">Invoice Amount</TableHead>
-                  </TableRow>
-                </TableHeader>
-                <TableBody>
-                  {paginated.map((grn) => (
-                    <GoodsReceiptRow
-                      key={grn.id}
-                      grn={grn}
-                      onClick={() =>
-                        router.push(
-                          routes.resources.goodsReceipts.detail(grn.id).href
-                        )
-                      }
-                    />
-                  ))}
-                </TableBody>
-              </Table>
-            </CardContent>
-            <div className="flex items-center justify-between border-t px-4 py-2">
-              <span className="text-sm text-zinc-500">
-                {startIndex + 1}–{endIndex} of {filtered.length} GRN
-                {filtered.length === 1 ? '' : 's'}
-              </span>
-              <Pagination
-                currentPage={currentPage}
-                totalPages={totalPages}
-                onPageChange={setCurrentPage}
-              />
-            </div>
-          </>
-        ) : (
-          <CardContent>
-            <Empty variant="default">
-              <EmptyMedia variant="icon">
-                <Receipt className="size-6" />
-              </EmptyMedia>
-              <EmptyHeader>
-                <EmptyTitle>No goods receipts found</EmptyTitle>
-                <EmptyDescription>
-                  {hasActiveFilters
-                    ? 'No GRNs match your search. Try adjusting your filters.'
-                    : 'Record your first GRN to get started.'}
-                </EmptyDescription>
-              </EmptyHeader>
-              {!hasActiveFilters && (
-                <Button asChild>
-                  <Link href={routes.resources.goodsReceipts.new}>
-                    Record GRN
-                  </Link>
-                </Button>
-              )}
-            </Empty>
-          </CardContent>
-        )}
-      </Card>
+      <GoodsReceiptTable
+        paginated={paginated}
+        filteredCount={filtered.length}
+        startIndex={startIndex}
+        itemsPerPage={itemsPerPage}
+        onItemsPerPageChange={(n) => {
+          setItemsPerPage(n);
+          setCurrentPage(1);
+        }}
+        currentPage={currentPage}
+        totalPages={totalPages}
+        onPageChange={setCurrentPage}
+        hasActiveFilters={hasActiveFilters}
+        searchValue={searchQuery}
+        onSearchChange={(v) => {
+          setSearchQuery(v);
+          setCurrentPage(1);
+        }}
+        projectFilter={projectFilter}
+        onProjectChange={(v) => {
+          setProjectFilter(v);
+          setCurrentPage(1);
+        }}
+        projectOptions={projectOptions}
+      />
     </div>
   );
 }
