+++ features/grn/components/index.ts
@@ -2,3 +2,5 @@ export * from './grn-alert-dialogs';
 export * from './grn-items-card';
 export * from './grn-receipt-info-card';
 export * from './grn-vendor-po-card';
+export { GoodsReceiptTable } from './goods-receipt-table';
+export { GoodsReceiptRow } from './goods-receipt-row';
